apps/db-exporter/requirements.txt
mysql-connector-python
psutil
prometheus-client
